Looks like I got bad info! You CAN have multiple checks in a single try block and it works quite nicely, actually.
However, I did run into some really weird logic problems that I still don't quite understand. Despite my conditional showing as False in the debugger, it would not enter the conditional even though the condition was set for "False." It was truly bizarre and I took way too long debugging to try to figure it out. Finally just rewrote the conditional check to check to make sure that numbers that should be negative are, indeed, negative, and it works now, so I really honestly don't know what it was on about.